Add border to tooltip required tech images:
The borders match the area of tech they are from. Anomalies(Events) have a black border. This is done during initialization of the tooltip but could be changed to occur every time the tooltip is displayed in case a tree isn't fully loaded. I haven't seen any issues with it being in initialization yet.