Major Update - Research node status.
Nodes have a X in lower right corner that can be clicked into a checkmark. An "active" node then refreshes the tree with connectors to children in the color of the area of research. Those nodes then can be clicked to further the research chain.

Additionally, the tooltips have been updated to show status of Required Technology with an X or checkmark next to the name.

Finally, since the areas of research are being used all over the place I put them in a variable and made it so they can be iterated in multiple places.